+++ NEWS.md
@@ -2,8 +2,9 @@
 
 ## rlistings 0.1.1
  * add title, subtitle, and (main and prov) footer support
- * now depends on dplyr instea dof magrittr to hopefully avoid var_labels droppage issues
+ * now depends on dplyr instead of magrittr to hopefully avoid var_labels droppage issues
  * `paginate_listing` now supports pagination in both directions
+ * Add regression tests for listings DSL01 and DSL02.
 
 ## rlistings 0.1.0
  * Initial experimental rlistings API. Everything subject to change.